v0.8.0: maintenance release
