+++ .github/workflows/python-publish.yml
@@ -7,7 +7,7 @@ on:
   # normal behavior: run when a new release is created
   release:
     types: [created]
-  # allow running manually
+  # allow running manually on main (restriction within job)
   workflow_dispatch:
 
 permissions:
@@ -22,18 +22,18 @@ jobs:
       url: https://pypi.org/project/piffle
     permissions:
       id-token: write  # IMPORTANT: this permission is mandatory for trusted publishing
-
+    if: github.event_name == 'release' || (github.event_name == 'workflow_dispatch' && github.ref == 'main')
     steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: "3.11"
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install build
-    - name: Build package
-      run: python -m build
-    - name: Publish package
-      uses: pypa/gh-action-pypi-publish@release/v1
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install build
+      - name: Build package
+        run: python -m build
+      - name: Publish package
+        uses: pypa/gh-action-pypi-publish@release/v1
