+++ .github/workflows/ruff-checks.yml
@@ -1,0 +1,30 @@
+name: Check Ruff linter + formatter
+on:
+  # Run on pull requests to develop or main
+  pull_request:
+    branches: [develop, main]
+  # Allow running manually
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+concurrency:
+  # Cancel existing job(s) for workflow when a new one is queued
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  ruff:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Run Ruff linter
+        uses: astral-sh/ruff-action@v3
+        with:
+          args: "check --output-format=github"
+      - name: Run Ruff formatter
+        # NOTE: ruff format does not currently support github output format
+        run: ruff format --check --diff
+        # Check formatting even if the previous step failed
+        if: always()
