+++ .github/workflows/unit-tests.yml
@@ -1,10 +1,19 @@
 name: unit tests
 
 on:
-  push: # run on every push or PR to any branch
+  push:  # run on all branches
   pull_request:
-  schedule: # run automatically on main branch each Tuesday at 11am
+    branches: [develop, main]
+  schedule:  # run automatically on main branch each Tuesday at 11am
     - cron: "0 16 * * 2"
+
+permissions:
+  contents: read
+
+concurrency:
+  # Cancel existing job(s) for workflow when a new one is queued
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
 
 jobs:
   python-unit:
@@ -12,41 +21,33 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python: ["3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
+        python: ["3.10", "3.11", "3.12", "3.13", "3.14"]
 
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
 
-      - name: Setup Python
-        uses: actions/setup-python@v4
+      - name: Install uv and Python version
+        uses: astral-sh/setup-uv@v7
         with:
           python-version: ${{ matrix.python }}
 
-      # We base the python cache on the hash of all requirements files, so that
-      # if any change, the cache is invalidated.
-      - name: Cache pip
-        uses: actions/cache@v3
-        with:
-          path: ~/.cache/pip
-          key: pip-${{ hashFiles('setup.py') }}
-          restore-keys: |
-            pip-${{ hashFiles('setup.py') }}
-            pip-
-
       - name: Install package with dependencies
-        run: |
-          pip install -e .
-          pip install -e '.[test]'
-          pip install codecov
+        run: uv sync --extra test
 
       - name: Run pytest
-        run: py.test --cov=piffle --cov-report=xml
+        run: uv run pytest --cov=piffle --cov=tests --cov-report=xml --cov-report=term
 
+      # Only upload test coverage for Python 3.12
       - name: Upload test coverage to Codecov
-        uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v5
+        if: matrix.python == '3.12'
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: ./coverage.xml
+          fail_ci_if_error: false
 
-  # Set the color of the slack message used in the next step based on the
+      # Set the color of the slack message used in the next step based on the
       # status of the build: "danger" for failure, "good" for success,
       # "warning" for error
       - name: Set Slack message color based on build status
@@ -68,4 +69,4 @@ jobs:
           SLACK_TITLE: "Workflow `${{ github.workflow }}` (python ${{ matrix.python }}): ${{ job.status }}"
           SLACK_MESSAGE: "Run <https://github.com/${{ github.repository }}/actions/runs/${{ github.run_id }}|#${{ github.run_number }}> on <https://github.com/${{ github.repository }}/|${{ github.repository }}@${{ github.ref }}>"
           SLACK_FOOTER: "<https://github.com/${{ github.repository }}/commit/${{ github.sha }}|View commit>"
-          MSG_MINIMAL: true # use compact slack message format
+          MSG_MINIMAL: true  # use compact slack message format
