+++ src/piffle/iiif_dataclasses/base.py
