+++ src/piffle/iiif_dataclasses/image2.py
