+++ src/piffle/iiif_dataclasses/image3.py
