+++ yamlfmt.yml
@@ -1,0 +1,7 @@
+# This is the config file for yamlfmt
+# For more details, see:
+#   https://github.com/google/yamlfmt/blob/main/docs/config-file.md
+formatter:
+  retain_line_breaks_single: true
+  pad_line_comments: 2
+  scan_folded_as_literal: true
